fix habit editing
